Set up all redesign files
