+++ .gitignore
@@ -5,4 +5,3 @@ cache
 build
 dist
 *.egg-info
-assets/theme.css
