+++ assets/__style_guide.css
@@ -1,0 +1,248 @@
+/*
+Copyright 2024 D-Wave
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+/* A custom base of style rules defined specifically for the demos */
+
+*, *::before, *::after {
+    box-sizing: border-box;
+}
+
+body {
+    color: var(--grey-darkest);
+    font-size: 1.125rem;
+    line-height: 1.5rem;
+    margin: 0;
+}
+
+h1, h2, h3, h4, h5, h6, td, th,
+span, a, p, label, button, input {
+    font-family: var(--font);
+    color: var(--blue-darkest);
+}
+
+h1, h2, h3, h4, h5, h6 {
+    font-weight: 600;
+    margin-top: 0;
+}
+
+h1 {
+    font-size: 2.5rem;
+    line-height: 3rem;
+    margin-bottom: 1rem;
+    font-weight: 400;
+}
+
+h2 {
+    font-size: 1.625rem;
+}
+
+h3 {
+    font-size: 1.5rem;
+}
+
+h4 {
+    font-size: 1.25rem;
+    margin-top: 0;
+    font-weight: 500;
+}
+
+h5 {
+    font-size: 1.125rem;
+    margin: 1rem 0;
+    font-weight: 600;
+    margin-top: 1rem;
+    margin-bottom: 1rem;
+}
+
+label {
+    margin: 1rem 0 0.5rem;
+    font-weight: 600;
+    display: block;
+}
+
+hr {
+    margin: 0.625rem 0;
+    border-width: 0;
+    border-top: 1px solid var(--grey-light);
+}
+
+table {
+    margin-bottom: 1.25rem;
+    font-size: 1rem;
+    border-collapse: collapse;
+    border: 1px solid var(--grey-light);
+}
+
+thead {
+    border-bottom: 1px solid var(--grey-light);
+    font-weight: 600;
+    background-color: var(--blue-darker);
+}
+
+tfoot {
+    border-top: 1px solid var(--grey-light);
+    font-weight: 600;
+}
+
+th {
+    font-weight: 600;
+    color: white;
+    padding: 0.5rem;
+    text-align: left;
+}
+
+td {
+    padding: 0.25rem 0.5rem;
+    text-align: left;
+}
+
+th:first-child, td:first-child {
+    padding-left: 0.75rem;
+}
+
+th:last-child, td:last-child {
+    padding-right: 0.75rem;
+}
+
+input[type="checkbox"],
+input[type="radio"] {
+    accent-color: var(--blue-darker);
+    margin: 0 0.5rem 0 0;
+}
+
+.display-none {
+    display: none !important;
+}
+  
+#app-container {
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+}
+
+.banner {
+    height: var(--banner-height);
+    background-color: #12131f;
+    padding: 0 2rem;
+    border-bottom: 2px solid var(--teal);
+
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.banner img {
+    height: 1rem;
+}
+
+.columns-main {
+    display: flex;
+    flex: 1;
+}
+
+.left-column {
+    display: flex;
+    height: 100%;
+    background-color: var(--grey-lighter);
+    box-shadow: var(--box-shadow);
+    z-index: 10;
+}
+
+.right-column {
+    width: 100%;
+    min-height: 40rem;
+}
+
+.title-section {
+    padding: 2rem 1rem 1.5rem;
+    background-color: var(--blue-darkest);
+    background-image: url("background.svg");
+    background-position: 30% 0;
+    height: var(--title-section-height);
+}
+
+.title-section h1 {
+    color: white;
+}
+
+.title-section p {
+    margin-bottom: 0;
+    color: white;
+}
+
+.settings {
+    margin: 1rem 0 2rem;
+}
+
+.form-section {
+    display: flex;
+    flex: 1;
+}
+
+.settings-and-buttons-wrapper {
+    padding: 0 1rem 2rem;
+    flex: 1;
+}
+
+/* The following rules are overwriting Dash styling */
+.VirtualizedSelectFocusedOption {
+    background-color: var(--grey-lighter);
+}
+
+.is-focused:not(.is-open) > .Select-control {
+    border-color: var(--teal);
+    box-shadow: none;
+}
+
+.Select {
+    font-family: var(--font);
+}
+
+div.dash-sk-circle {
+    height: 4rem;
+    width: 4rem;
+}
+
+/* Hide the skip link off-screen by default */
+.skip-link {
+    position: absolute;
+    left: -1.5rem;
+    width: 1px;
+    height: 1px;
+    overflow: hidden;
+}
+
+/* Make the skip link visible when it receives keyboard focus */
+.skip-link:focus {
+    color: var(--teal);
+    position: static;
+    width: auto;
+    height: auto;
+    overflow: visible;
+    display: inline-block;
+    padding: 1rem;
+    z-index: 1000;
+}
+
+@media screen and (max-width: 1000px) {
+    .columns-main {
+        flex-direction: column;
+    }
+
+    .form-section {
+        flex-direction: column;
+    }
+}
