+++ assets/_buttons.css
@@ -1,0 +1,76 @@
+/*
+Copyright 2024 D-Wave
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+/* Style rules for buttons */
+
+button {
+    border: none;
+    font-size: 0.875rem;
+    font-weight: 600;
+    line-height: 1.75;
+    border-radius: 4px;
+    padding: 0.75rem 1.5rem;
+    cursor: pointer;
+    transition: all 0.2s ease-in-out;
+}
+
+.button {
+    font-size: 0.875rem;
+    font-weight: 600;
+    line-height: 1.75;
+    border-radius: 4px;
+    padding: 0.75rem 1.5rem;
+    color: white;
+    text-transform: uppercase;
+    background: radial-gradient(61.22% 95.86% at 26.96% 100%, #4c71c6 0%, var(--blue-darker) 100%);
+    transition: background-color 250ms cubic-bezier(0.4, 0, 0.2, 1),
+                box-shadow 250ms cubic-bezier(0.4, 0, 0.2, 1),
+                border-color 250ms cubic-bezier(0.4, 0, 0.2, 1);
+    box-shadow: rgba(29, 30, 36, 0.19) 0px 5px 14px 0px;
+    border: 1px solid rgba(76, 113, 198, 0.4);
+}
+
+.button:hover {
+    background: radial-gradient(61.22% 95.86% at 26.96% 100%, var(--blue-darker) 0%, var(--blue-darker) 100%);
+}
+
+#run-button,
+#cancel-button {
+    width: 100%;
+}
+
+#cancel-button {
+    background: radial-gradient(61.22% 95.86% at 26.96% 100%, var(--red-light) 0%, var(--red-dark) 100%);
+    border: 1px solid rgb(245, 118, 119, 0.4);
+}
+
+#cancel-button:hover {
+    background: radial-gradient(61.22% 95.86% at 26.96% 100%, var(--red-dark) 0%, var(--red-dark) 100%);
+}
+
+button:disabled {
+    cursor: not-allowed;
+    filter: brightness(70%);
+    opacity: 0.7;
+}
+
+button:focus-visible {
+    box-shadow: 0 0 0 4px var(--teal);
+}
+
+.button-small {
+    padding: 0.35rem 1rem;
+}
