+++ assets/_collapse.css
@@ -1,0 +1,186 @@
+/*
+Copyright 2024 D-Wave
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+/* Style rules for collapsible dropdowns like left-column-collapse and problem-details-collapse */
+
+.collapse-arrow {
+    height: 1.5rem;
+    position: relative;
+    width: 1.4rem;
+    transition: border-color 0.25s ease-in-out;
+}
+
+.collapse-arrow:before,
+.collapse-arrow:after {
+    content: "";
+    display: block;
+    position: absolute;
+    height: 0.25rem;
+    width: 1.5rem;
+    background: var(--blue-darker);
+    border-radius: 0.125rem;
+}
+
+.collapse-arrow:before {
+    rotate: -70deg;
+    top: 0;
+}
+
+.collapse-arrow:after {
+    rotate: 70deg;
+    bottom: 0;
+}
+
+.settings-and-buttons-wrapper {
+    height: calc(100vh - var(--title-section-height));
+    overflow-y: auto;
+    direction: rtl;
+}
+
+.settings-and-buttons {
+    direction: ltr;
+}
+
+.left-column .left-column-layer-1 {
+    width: var(--left-col-width);
+    transition: width 0.5s ease-in-out;
+    direction: rtl;
+}
+
+.details-to-collapse {
+    height: var(--problem-details-height);
+    transition: height 0.5s ease-in-out;
+    overflow: hidden;
+}
+
+.left-column .left-column-layer-2 {
+    width: var(--left-col-width);
+    box-sizing: border-box;
+    direction: ltr;
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+}
+
+.details-collapse-wrapper {
+    margin-bottom: 1.25rem;
+    overflow: hidden;
+}
+
+.left-column-collapse {
+    background: white;
+    border: none;
+    height: 100%;
+    border-radius: 0;
+    padding: 0;
+    display: block;
+}
+
+.left-column-collapse:hover,
+.left-column-collapse:focus {
+    background: var(--grey-lighter);
+}
+
+.details-collapse,
+.details-collapse:hover,
+.details-collapse:focus {
+    background: none;
+    display: flex;
+    align-items: center;
+    padding: 0 1.25rem 0 0;
+    text-transform: none;
+    color: var(--blue-dark);
+}
+
+.details-collapse .collapse-arrow {
+    transform: rotate(90deg);
+    width: 3rem;
+    margin-top: 1.5rem;
+    margin-left: 0.5rem;
+    color: var(--blue-darkest);
+}
+
+.left-column-collapse:hover .collapse-arrow {
+    border-color: var(--grey);
+}
+
+.collapsed .left-column-collapse {
+    padding: 0 0.25rem 0 0;
+}
+
+.collapsed .left-column-collapse .collapse-arrow {
+    margin-right: -4px;
+    transform: rotate(180deg);
+}
+
+.title-section {
+    transition: all 0.5s ease-in-out;
+}
+
+.collapsed .title-section {
+    height: 0;
+    padding: 0;
+    overflow: hidden;
+}
+
+.collapsed .details-collapse .collapse-arrow {
+    margin-bottom: 1.5rem;
+    margin-top: 0;
+    transform: rotate(-90deg);
+}
+
+.collapsed .details-to-collapse {
+    height: 0;
+}
+
+.collapsed .left-column-layer-1 {
+    width: 1.5rem;
+}
+
+@media screen and (max-width: 1000px) {
+    .left-column {
+        height: auto;
+        width: 100%;
+    }
+
+    .left-column .left-column-layer-1,
+    .left-column .left-column-layer-2,
+    .left-column .left-column-collapse {
+        width: 100%;
+    }
+
+    .left-column-collapse .collapse-arrow {
+        transform: rotate(90deg);
+        margin: auto;
+    }
+
+    .collapsed .title-section {
+        height: 0;
+        overflow-y: hidden;
+    }
+
+    .collapsed .settings-and-buttons-wrapper {
+        height: 0;
+        flex: unset;
+        padding: 0;
+        overflow-y: hidden;
+    }
+
+    .collapsed .left-column-collapse .collapse-arrow {
+        margin-right: auto;
+        transform: rotate(-90deg);
+    }
+}
