--- assets/job_shop.css
@@ -1,354 +1,0 @@
-/* 
-This file is forked from apps/dash-clinical-analytics/assets/clinical-analytics.css under the following license
-
-MIT License
-
-Copyright (c) 2019 Plotly
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-
-Modifications are licensed under
-
-Apache License, Version 2.0
-(see ./LICENSE for details)
-
-*/
-
-
-/* theme.css is automatically generated, see app.py for more details */
-@import url("theme.css");
-
-:root {
-    --blue-bright: #03B8FF;
-    --blue-light: #2A7DE1;
-    --blue-dark: #074C91;
-    --red-light: #C70039;
-    --red-dark: #900C3F;
-    --orange: #FF7006;
-    --teal-light: #06ECDC;
-    --teal-dark: #17BEBB;
-    --teal-darker: #008C82;
-    --grey-lighter: #EEEEEE;
-    --grey-light: #DDDDDD;
-    --grey: #AAAAAA;
-    --grey-medium: #777777;
-    --grey-dark: #222222;
-    --banner-height: 5.5rem;
-    --left-col-width: 35rem;
-}
-
-body {
-    background-color: #f9f9f9;
-    color: var(--grey-dark);
-    font-size: 1.4rem;
-    margin: 0;
-}
-
-.display-none {
-    display: none;
-}
-
-#banner {
-    flex: 0 0 var(--banner-height);
-    background-color: var(--theme);
-
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-}
-
-#banner img {
-    position: relative;
-    height: 3rem;
-    width: auto;
-    margin-left: 2rem;
-}
-
-
-h1, h2, h3, h4, h5, h6, td, th, span, a, p, label {
-    font-family: "Helvetica Neue", sans-serif;
-}
-
-h1, h2, h3, h4, h5, h6 {
-    color: var(--theme);
-    font-weight: 400;
-}
-
-h1 {
-    font-size: 3rem;
-    margin-top: 3rem;
-}
-
-h3 {
-    font-size: 2.4rem;
-    margin: 2.5rem 0 1rem;
-}
-
-h3.gantt-title {
-    margin-bottom: 0;
-    margin-left: 4.6rem;
-}
-
-h4 {
-    font-size: 2rem;
-    margin-top: 2rem;
-    font-weight: 500;
-}
-
-h5 {
-    font-size: 1.8rem;
-    font-weight: 600;
-}
-
-label {
-    margin: 2rem 0 0.2rem;
-    font-weight: 500;
-}
-
-#solver-select label {
-    margin: 0;
-    font-weight: 400;
-    display: flex;
-}
-
-#solver-select.hide-classic label:last-child {
-    display: none !important;
-}
-
-input[type="checkbox"] {
-    margin: 0 1rem 0 0;
-}
-
-#app-container {
-    min-width: 1000px;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-}
-
-#columns {
-    display: flex;
-    flex: 1;
-}
-
-#left-column {
-    display: flex;
-}
-
-#left-column > div > div {
-    padding: 0 2rem;
-    width: calc(var(--left-col-width) - 4rem);
-}
-
-#left-column > div:first-child {
-    width: var(--left-col-width);
-    transition: width 0.6s ease-in-out;
-    overflow: hidden;
-}
-
-#left-column.collapsed > div:first-child {
-    width: 0;
-}
-
-#right-column {
-    background-color: var(--grey-light);
-    padding: 2rem 0 0;
-    width: 100%;
-}
-
-.table-row {
-    height: 63px;
-    vertical-align: middle;
-    font-size: 13px;
-}
-
-.wait_time_graph > div, .patient_score_graph > div {
-    height: 100% !important;
-    width: 100% !important;
-}
-
-.draglayer {
-    cursor: pointer !important;
-}
-
-#control-card > p {
-    font-weight: bold;
-}
-
-#gantt_chart_card, #wait_time_card {
-    background-color: #FFFFFF;
-    padding: 1rem;
-    margin: auto;
-    margin-top: 20px;
-}
-
-#gantt_chart_card > b, #wait_time_card > b {
-    line-height: 2.5;
-}
-
-
-#gantt_chart {
-    height: 400px;
-}
-
-.table-row:nth-child(even) {
-    background-color: #f2f2f2;
-}
-
-
-#header_wait_time_min, #header_care_score {
-    display: table;
-}
-
-#header_wait_time_min > b, #header_care_score > b {
-    display: table-cell;
-    vertical-align: middle;
-}
-
-.tab-container {
-    border-bottom: 3px solid var(--theme-secondary);
-    flex-direction: row !important;
-}
-
-.tab {
-    margin: 0 2rem;
-    border-top-left-radius: 0.6rem;
-    border-top-right-radius: 0.6rem;
-    width: calc(100% / 2) !important;
-}
-
-div.tab.tab--selected {
-    border: 3px solid var(--theme-secondary) !important;
-    border-bottom: none !important;
-    cursor: default !important;
-    box-shadow: 0 6px 0 -3px white;
-}
-
-div.tab.tab--disabled {
-    cursor: not-allowed !important;
-}
-
-.tab:first-child {
-    margin-right: 1rem;
-}
-
-.tab:last-child {
-    margin-left: 1rem;
-}
-
-.tab-content,
-.tab-parent {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-}
-
-.tab-content {
-    background-color: white;
-}
-
-.tab-content > div {
-    background-color: white;
-    padding: 1rem 1.5rem 3rem;
-}
-
-#run-button,
-#cancel-button {
-    width: 100%;
-    font-size: 1.4rem;
-    line-height: 1.4rem;
-    padding: 1.8rem;
-    height: auto;
-    color: white;
-    transition: all 0.2s ease-in-out;
-    margin-bottom: 4rem;
-    border: none;
-}
-
-#left-column-collapse {
-    background: white;
-    border: 1px var(--grey-lighter);
-    border-style: none solid none none;
-    height: 100%;
-    border-radius: 0;
-    padding: 0 0 0 0.5rem;
-}
-
-#left-column-collapse div {
-    border-right: 4px solid var(--grey-light);
-    border-bottom: 4px solid var(--grey-light);
-    transform: rotate(135deg) skew(165deg, 165deg);
-    height: 2rem;
-    width: 2rem;
-    margin-right: -0.3rem;
-    transition: border-color 0.25s ease-in-out;
-}
-
-#left-column-collapse:hover div {
-    border-color: var(--grey);
-}
-
-.collapsed #left-column-collapse {
-    padding: 0 0.5rem 0 0;
-}
-
-.collapsed #left-column-collapse div {
-    transform: rotate(315deg) skew(160deg, 160deg);
-}
-
-.VirtualizedSelectFocusedOption {
-    background-color: var(--grey-lighter);
-}
-
-.is-focused:not(.is-open)>.Select-control {
-    border-color: var(--theme-secondary);
-    box-shadow: none;
-}
-
-#run-button {
-    background-color: var(--theme);
-}
-
-#run-button:hover {
-    filter: brightness(80%);
-}
-
-#cancel-button {
-    background-color: var(--red-light);
-}
-
-#cancel-button:hover {
-    background-color: var(--red-dark);
-}
-
-.tab-fail {
-    border-top: 3px solid var(--red-light) !important;
-}
-
-.tab-success {
-    border-top: 3px solid var(--teal-darker) !important;
-}
-
-.tab-warning {
-    border-top: 3px solid var(--orange) !important;
-}
-
-#solver-time-limit {
-    margin-bottom: 20px;
-}
